// components/Organisms/CalmInfoHome/types.ts
export interface IPosts {
    firstImage: string
    firstImageMobile?: string
    firstTitle: string
    firstText: string
    firstButtonLink: string
    firstButtonText: string
    secondImage: string
    secondImageMobile?: string
    secondTitle: string
    secondText: string
    secondButtonLink: string
    secondButtonText: string
  }
  
 export interface IProps {
    info?: IPosts
  }
  